ccs: refactor ccs circuit in ir
