+++ src/ccs/ir/mod.rs
@@ -1,8 +1,16 @@
-use std::{collections::HashMap, fmt::Debug};
-
 use crate::{field::Field, poly::Expr, util::UUID};
+use std::{collections::HashMap, fmt::Debug, hash::Hash};
 pub mod assignments;
 pub mod circuit;
+
+use assignments::Assignments;
+use circuit::CCSCircuit;
+
+pub type MatrixCoeffsAndOffset<F> = Vec<Vec<(CoeffsForProds<F>, usize)>>;
+
+// Use to mark the public inputs
+// ((step_idx, step_UUID), assignment_offset)
+pub type AssignmentOffsets = HashMap<(usize, u128), usize>;
 
 #[derive(Clone)]
 pub struct Poly<F> {
@@ -18,7 +26,7 @@ impl<F: Debug> Debug for Poly<F> {
     }
 }
 
-// (signal_uuid, step_uuid, annotation, pos)
+// (signal_uuid, step_uuid, annotation, is_next)
 pub type PolyExpr<F> = Expr<F, (UUID, UUID, String, bool)>;
 
 pub type Coeff<F> = (F, UUID, bool);
@@ -173,4 +181,133 @@ fn factor_mul<F: Clone + Field>(polys: Vec<PolyExpr<F>>) -> PolyExpr<F> {
     }
 
     new_polys
+}
+
+// Circuit
+// matrix_coeffs_and_offsets: Temporarily record all circuit coefficients in each step
+//                          (It is not possible to build a complete circuit until the trace is
+// traced,                          so it is necessary to temporarily record the circuit structure
+// in each step.) exposed: All the offsets of public inputs
+// witness: All the offsets of the witnesses
+// ccs: The ccs circuit
+#[derive(Debug)]
+pub struct Circuit<F> {
+    pub ast_id: UUID,
+
+    pub matrix_coeffs_and_offsets: HashMap<UUID, MatrixCoeffsAndOffset<F>>,
+    pub exposed: Vec<(usize, UUID)>,
+    pub witness: HashMap<UUID, Vec<UUID>>,
+
+    pub ccs: CCSCircuit<F>,
+}
+
+impl<F: Field + From<u64> + Hash> Circuit<F> {
+    pub fn instance(&self, assignments: &Option<Assignments<F>>) -> HashMap<(usize, UUID), F> {
+        let mut exposes = HashMap::new();
+        if !self.exposed.is_empty() {
+            for (step_idx, id) in self.exposed.iter() {
+                if let Some(witness) = assignments {
+                    exposes.insert((*step_idx, *id), witness.get(*step_idx, *id));
+                }
+            }
+        }
+        exposes
+    }
+}
+#[cfg(test)]
+mod tests {
+    use std::vec;
+
+    use super::{assignments::Z, circuit::Matrix, *};
+    use halo2_proofs::halo2curves::bn256::Fr;
+
+    fn create_ccs_circuit<F: Field + From<u64> + Hash>(
+        n: usize,
+        m: usize,
+        l: usize,
+        matrices: &[Vec<(usize, usize, F)>],
+        selectors: &[Vec<(usize, F)>],
+        constants: &[F],
+    ) -> CCSCircuit<F> {
+        let mut degree = 0;
+        assert_eq!(selectors.len(), constants.len());
+        for selector in selectors.iter() {
+            for &(s, _) in selector {
+                assert!(s < matrices.len())
+            }
+            degree = degree.max(selector.len())
+        }
+        let matrices: Vec<Matrix<F>> = matrices
+            .iter()
+            .map(|cells| Matrix::from_values(n, m, cells))
+            .collect();
+
+        CCSCircuit {
+            n,
+            m,
+            t: matrices.len(),
+            q: constants.len(),
+            l,
+            d: degree,
+            constants: constants.to_vec(),
+            selectors: selectors.to_vec(),
+            matrices,
+        }
+    }
+
+    #[test]
+    fn test_ccs() {
+        let n = 7;
+        let m = 4;
+        let l = 3;
+
+        let m0 = vec![
+            (0, 1, Fr::ONE),
+            (1, 2, Fr::ONE),
+            (2, 3, Fr::ONE),
+            (3, 6, Fr::ONE),
+        ];
+        let m1 = vec![
+            (0, 1, Fr::ONE),
+            (1, 2, Fr::ONE),
+            (2, 4, Fr::ONE),
+            (3, 6, Fr::ONE),
+        ];
+        let m2 = vec![
+            (0, 1, Fr::ONE),
+            (1, 2, Fr::ONE),
+            (2, 5, Fr::ONE),
+            (3, 6, Fr::ONE),
+        ];
+        let m3 = vec![(0, 0, Fr::ONE), (1, 0, Fr::ONE)];
+        let m4 = vec![(2, 0, Fr::from(2))];
+        let m5 = vec![(2, 0, Fr::from(2))];
+        let m6 = vec![
+            (0, 0, Fr::ONE.neg()),
+            (1, 0, Fr::ONE.neg()),
+            (2, 0, Fr::ONE.neg()),
+        ];
+        let m7 = vec![(0, 0, Fr::ZERO)];
+        let matrices = vec![m0, m1, m2, m3, m4, m5, m6, m7];
+
+        let selectors = vec![
+            vec![(3, Fr::ONE), (0, Fr::ONE), (1, Fr::ONE)],
+            vec![(4, Fr::ONE), (0, Fr::ONE)],
+            vec![(5, Fr::ONE), (1, Fr::ONE)],
+            vec![(6, Fr::ONE), (2, Fr::ONE)],
+            vec![(7, Fr::ONE)],
+        ];
+        let constants: Vec<Fr> = vec![Fr::ONE, Fr::ONE, Fr::ONE, Fr::ONE, Fr::ONE];
+
+        let ccs_circuit = create_ccs_circuit(n, m, l, &matrices, &selectors, &constants);
+
+        let z = Z::from_values(
+            &[Fr::ZERO, Fr::ONE, Fr::from(2)],
+            &[Fr::from(3), Fr::from(10), Fr::from(43)],
+        );
+
+        let result = ccs_circuit.is_satisfied(&z);
+
+        println!("result = {}", result);
+    }
 }
